moves the components to shared location
