+++ package.json
@@ -50,6 +50,7 @@
     "@babel/helper-create-class-features-plugin": "^7.21.8",
     "@hmcts/ccpay-web-component": "6.0.13",
     "@hmcts/cookie-manager": "^1.0.0",
+    "@hmcts/frontend": "0.0.50-alpha",
     "@hmcts/nodejs-healthcheck": "^1.8.0",
     "@hmcts/nodejs-logging": "^4.0.4",
     "@hmcts/properties-volume": "^1.0.0",
@@ -63,6 +64,7 @@
     "ejs": "^3.1.7",
     "express": "^4.16.4",
     "express-session": "^1.17.1",
+    "govuk-frontend": "^3.12.0",
     "helmet": "^3.22.0",
     "http-status-codes": "^1.3.0",
     "i": "^0.3.7",
