+++ src/app/app.component.html
@@ -16,19 +16,4 @@ name="paybubble"
 
 </div>
 
-<footer class="group js-footer" id="footer" role="contentinfo">
-  <div class="footer-wrapper">
-    <div class="footer-meta">
-      <div class="footer-meta-inner">
-        <div class="open-government-licence">
-          <p class="logo">
-            <a href="https://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/" rel="license">Open Government Licence</a>
-          </p>
-          <p>All content is available under the
-            <a href="https://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/"
-              rel="license">Open Government Licence v3.0</a>, except where otherwise stated</p>
-        </div>
-      </div>
-    </div>
-  </div>
-</footer>
+<ccpay-bubble-app-footer></ccpay-bubble-app-footer>
