+++ src/app/app.constants.ts
@@ -1,0 +1,14 @@
+
+const footerDataNavigation = {
+    items: [
+      { text: 'Cookies', href: '/cookies', target: '_blank' },
+      { text: 'Accessibility Statement', href: '/accessibility', target: '_blank' },
+      { text: 'Privacy policy', href: 'https://hmcts-access.service.gov.uk/privacy-policy', target: '_blank' },
+      { text: 'Terms and conditions', href: '/https://hmcts-access.service.gov.uk/terms-and-conditions', target: '_blank' },
+      { text: 'Contact us', href: 'https://hmcts-access.service.gov.uk/contact-us', target: '_blank' }
+    ]
+};
+
+export class AppConstants {
+    public static FOOTER_DATA_NAVIGATION = footerDataNavigation;
+}
