+++ src/app/app.module.ts
@@ -38,6 +38,8 @@ import { CookiePolicyComponent } from './components/cookie-policy/cookie-policy.
 import { CookieDetailsComponent } from './components/cookie-details/cookie-details.component';
 import { windowProvider, windowToken } from '../window';
 import { CookieTableComponent } from './components/cookie-table/cookie-table.component';
+import { HmctsGlobalFooterComponent } from './shared/components/hmcts-global-footer/hmcts-global-footer.component';
+import { FooterComponent } from './shared/components/footer/footer.component';
 
 const nonProductionProviders = [{
   provide: HTTP_INTERCEPTORS,
@@ -65,7 +67,9 @@ const nonProductionProviders = [{
     CookieBannerComponent,
     CookiePolicyComponent,
     CookieDetailsComponent,
-    CookieTableComponent
+    CookieTableComponent,
+    HmctsGlobalFooterComponent,
+    FooterComponent
     ],
   imports: [
     BrowserModule,
