+++ src/app/components/ccd-search/ccd-search.component.html
@@ -123,7 +123,4 @@
     <button class="button" type="submit" (click)="searchFees()">Search</button>
     <div class="nocase-message" *ngIf="noCaseFoundInCCD">No matching cases found</div>
   </div>
-  
-  
-  
-  </form>
+</form>
