--- src/app/containers/footer/footer.component.html
@@ -1,3 +1,0 @@
-<exui-app-hmcts-global-footer>
-    
-</exui-app-hmcts-global-footer>
