+++ src/app/shared/components/footer/footer.component.html
@@ -1,0 +1,4 @@
+<ccpay-bubble-hmcts-global-footer
+[navigation]="navigationData"
+>
+</ccpay-bubble-hmcts-global-footer>
