+++ src/app/shared/components/footer/footer.component.ts
@@ -1,12 +1,16 @@
 import { Component, OnInit } from '@angular/core';
+import { AppConstants } from '../../../app.constants';
+import { Navigation } from './footer.model';
 
 @Component({
-  selector: 'exui-app-footer',
+  selector: 'ccpay-bubble-app-footer',
   templateUrl: './footer.component.html',
   styleUrls: ['./footer.component.scss']
 })
 
 export class FooterComponent implements OnInit {
+  public navigationData: Navigation = AppConstants.FOOTER_DATA_NAVIGATION;
+  
   constructor() {}
 
   public ngOnInit() {}
