+++ src/app/shared/components/footer/footer.model.ts
