+++ src/assets/stylesheets/main.scss
@@ -6,6 +6,7 @@ $path: '/public/images/';
 @import '../../../node_modules/govuk_template_jinja/assets/stylesheets/govuk-template.css';
 @import '../../../node_modules/govuk_template_jinja/assets/stylesheets/fonts.css';
 @import '../../../node_modules/govuk-frontend/all';
+// @import '../../../node_modules/@hmcts/frontend/all';
 
 
 body {
